Auto skill balancer
Can automatically skill balance teams or provide suggestions, both based on
threshold values.

We need a way to automatically trigger the skill balancer based on when players
might perceive the game as unfairly balanced.

To do this we calculate team contribution (player's worth to team as
kill-deaths) and take this averaged for the top players in each team (because
the top players affect how the game is perceived by the opposing team).

The difference in this avg values between the two teams determines how the game
is progressing (specifically how much a team is dominating TDM). This is
DATCPM - Difference in Average Team Contribution Per Minute.

We provide 4 modes of operation:

1) Do nothing :-)
2) Suggest - team strength is shown, with a suggestion to shuffle if
   DATCPM is higher than threshold.
3) Auto-min - team strength is shown, perform minmove for same reason
4) Auto-skuffle - team strength is shown, perform skuffle for same reason

The threshold is anecdotal and can be tweaked in the settings, though this will
leave eminent Swedish scientisticians unhappy :P

A new command !paautoskuffle (!ask) makes it possible to switch between the
modes in-game.

TODO:
 - rename !pabalance (!bal) to !pacheckbalance (!chk)
 - rename !paminmoves (!min) to !pabalance (!bal)