version 1.1 of the skill balancer
Release Notes for ver 1.1

With ver 1.0 we introduced DATCPM - Difference in Average Team Contribution Per
Minute. Our approach was correct but we were looking at small slices of data.
This works very well when there is a kill spree, but does nto work so well when
game play is slower. The new version reverts to using kill ratio, which has
been around since the very first version of the skill balancer.

In this version we calculate the difference in the average kill ratio of the
top players (for more than 3 player). If this is more than 0.5 then teams are
unfair. !Ask starts advising at a threshold difference of 0.38, and
progressively provides stronger descriptions at every increment of 0.1. If the
ratio difference is more than 1.0 the teams are hopeless.

We consider !bal to be stable enough to be a substitute for !teams, and can be
provided to regular players. Running !ask with autobalance (!ask 2) will keep
teams automatically balanced without moving too many players.